+++ packages/switch/src/style/index.ts
@@ -2,7 +2,7 @@ import { createStyles } from 'antd-style';
 import type { SwitchProps } from '../type';
 import { genCompoentStyle } from './genComponentStyle';
 import { DefaultTheme } from './theme';
-import { myDeepMerge } from '@yuxi-ui/util';
+import { myDeepMerge, tokenToCSSVar } from '@yuxi-ui/util';
 type SwitchPropsStyle = Pick<SwitchProps, 'prefixCls'>;
 export const useSwitchStyles = createStyles<SwitchPropsStyle>(
   ({ css }, props: SwitchPropsStyle) => {
@@ -19,32 +19,3 @@ export const useSwitchStyles = createStyles<SwitchPropsStyle>(
     });
   },
 );
-
-export type stringifield<T> = {
-  [k in keyof T]: string;
-};
-export const tokenToCSSVar = <T extends Record<string, any>>(
-  token: T,
-  prefixCls: string,
-) => {
-  const themeObject: Record<string, string | number> = {};
-  // 避免number类型期望，都转为string
-  const tokenVarMap = {} as stringifield<T>;
-
-  for (const key in token) {
-    if (Object.prototype.hasOwnProperty.call(token, key)) {
-      const value = token[key];
-      const kebabKey = camelToKebab(key);
-      const varName = `--${prefixCls}-${kebabKey}`;
-      tokenVarMap[key] = `var(${varName})`;
-      themeObject[varName] = value;
-    }
-  }
-  return { themeObject, tokenVarMap };
-};
-const camelToKebab = (val: string) => {
-  return val
-    .replace(/([a-z])([A-Z])/g, '$1-$2') // Insert hyphen between lowercase and uppercase
-    .replace(/([A-Z])([A-Z][a-z])/g, '$1-$2') // Handle consecutive capitals (like "HTTPRequest")
-    .toLowerCase();
-};