+++ packages/ui-library/src/index.ts
@@ -1,5 +1,5 @@
-// export * from '@yuxi-ui/config-provider';
+export * from '@yuxi-ui/config-provider';
 
-// export * from '@yuxi-ui/switch';
+export * from '@yuxi-ui/switch';
 
 export * from '@yuxi-ui/button';